+++ back/composer.json
@@ -42,7 +42,7 @@
         "phpunit/phpunit": "^9.5",
         "rector/rector": "^0.15.13",
         "symfony/browser-kit": "6.2.*",
-        "symfony/css-selector": "5.4.*",
+        "symfony/css-selector": "6.2.*",
         "symfony/phpunit-bridge": "^6.0",
         "symfony/stopwatch": "6.2.*",
         "symfony/web-profiler-bundle": "6.2.*",
