+++ back/config/bundles.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 return [
     Symfony\Bundle\FrameworkBundle\FrameworkBundle::class => ['all' => true],
