+++ back/config/packages/security.yaml
@@ -31,4 +31,4 @@ security:
     # Easy way to control access for large sections of your site
     # Note: Only the *first* access control that matches will be used
     access_control:
-        - { path: ^/test, roles: ROLE_USER }
+        - { path: ^/newsletter, roles: ROLE_USER }
