+++ back/phpunit.xml.dist
@@ -28,11 +28,6 @@
         <include>
             <directory suffix=".php">src</directory>
         </include>
-        <exclude>
-            <directory suffix=".php">Enum</directory>
-            <directory suffix=".php">Exception</directory>
-            <directory suffix=".php">Command</directory>
-        </exclude>
         <report>
             <clover outputFile="coverage/coverage.xml"/>
             <html outputDirectory="coverage"/>
@@ -41,6 +36,8 @@
             <file>src/Kernel.php</file>
             <directory>src/Controller</directory>
             <directory>src/Command</directory>
+            <directory>Enum</directory>
+            <directory>Exception</directory>
         </exclude>
     </coverage>
