+++ back/src/Controller/Newsletter/ContentController.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Controller;
+namespace App\Controller\Newsletter;
 
 use App\Service\NewsletterManager\Manager;
 use App\Service\NewsletterManager\NewspaperCreator;
@@ -12,9 +12,9 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
-class TestMailController extends AbstractController
+class ContentController extends AbstractController
 {
-    #[Route(path: '/test/mail/show', name: 'test_mail_show', methods: ['GET'])]
+    #[Route(path: '/newsletter/content/show', name: 'newsletter_content_show', methods: ['GET'])]
     public function show(
         Request $request,
         Manager $manager
@@ -32,19 +32,20 @@ class TestMailController extends AbstractController
         );
     }
 
-    #[Route(path: '/test/mail/all', name: 'all_mail_show', methods: ['GET'])]
+    #[Route(path: '/newsletter/content/all', name: 'newsletter_content_all', methods: ['GET'])]
     public function all(
         NewspaperCreator $creator,
         NewspaperRenderer $renderer
     ): Response {
         return new Response(
             $renderer->renderHtml(
-                $creator->createAllContent()
+                $creator->createAllContent(),
+                true
             )
         );
     }
 
-    #[Route(path: '/test/mail/one/{blockType}', name: 'test_mail_one', methods: ['GET'])]
+    #[Route(path: '/newsletter/content/one/{blockType}', name: 'newsletter_content_one', methods: ['GET'])]
     public function one(
         NewspaperCreator $creator,
         NewspaperRenderer $renderer,
