+++ back/tests/Functionnal/Controller/Newsletter/ContentControllerTest.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Tests\Functionnal\Controller;
+namespace App\Tests\Functionnal\Controller\Newsletter;
 
 use App\Tests\Functionnal\SetUserTrait;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
@@ -9,7 +9,7 @@ use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 /**
  * @internal
  */
-final class TestMailControllerTest extends WebTestCase
+final class ContentControllerTest extends WebTestCase
 {
     use SetUserTrait;
 
@@ -22,7 +22,7 @@ final class TestMailControllerTest extends WebTestCase
         $this->loginUser($client);
 
         // Request a specific page
-        $client->request('GET', '/test/mail/show/?date=2022-01-03');
+        $client->request('GET', '/newsletter/content/show/?date=2022-01-03');
         $this->assertSelectorTextContains('h1.test-img-random', 'Images de');
 
         // Validate a successful response and some content
@@ -38,7 +38,7 @@ final class TestMailControllerTest extends WebTestCase
         $this->loginUser($client);
 
         // Request a specific page
-        $client->request('GET', '/test/mail/show/?date=2022-01-02');
+        $client->request('GET', '/newsletter/content/show/?date=2022-01-02');
         $this->assertResponseIsSuccessful();
         $this->assertSelectorTextContains('h1.test-img-random', 'Images de');
 
@@ -55,7 +55,7 @@ final class TestMailControllerTest extends WebTestCase
         $this->loginUser($client);
 
         // Request a specific page
-        $client->request('GET', '/test/mail/show/?date=2022-01-01');
+        $client->request('GET', '/newsletter/content/show/?date=2022-01-01');
         $this->assertResponseIsSuccessful();
         $this->assertSelectorExists('h2#test-see-again');
 
@@ -72,7 +72,7 @@ final class TestMailControllerTest extends WebTestCase
         $this->loginUser($client);
 
         // Request a specific page
-        $client->request('GET', '/test/mail/all');
+        $client->request('GET', '/newsletter/content/all');
         $this->assertResponseIsSuccessful();
         $this->assertSelectorTextContains('h1.test-img-random', 'Images de');
 
@@ -89,7 +89,7 @@ final class TestMailControllerTest extends WebTestCase
         $this->loginUser($client);
 
         // Request a specific page
-        $client->request('GET', '/test/mail/show/?date=2022-01-04&force_twig=true');
+        $client->request('GET', '/newsletter/content/show/?date=2022-01-04&force_twig=true');
         $this->assertResponseIsSuccessful();
         $this->assertSelectorExists('h2#test-see-again');
